# docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"YantraMate API","version":"1.0.0","description":"REST API documentation for YantraMate"},"servers":[{"url":"http://localhost:3011","description":"Development server"},{"url":"https://z0n76c1j-3011.usw3.devtunnels.ms","description":"Production server"}],"paths":{"/api/v1/ask":{"post":{"summary":"Ask a question. Returns an answer with citations to source files and locators.","tags":["Ask"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["question"],"properties":{"question":{"type":"string","description":"The natural language question."},"projectId":{"type":"string","format":"uuid","description":"Optional project scope."},"topK":{"type":"integer","minimum":1,"maximum":20,"description":"Optional source cap before prompt planning."},"sessionId":{"type":"string","format":"uuid","description":"Optional session ID for multi-turn conversation context."}}}}}},"responses":{"200":{"description":"Answer with cited sources.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"answer":{"type":"string"},"sources":{"type":"array","items":{"type":"object","properties":{"chunkId":{"type":"string","format":"uuid"},"fileId":{"type":"string","format":"uuid"},"driveFileId":{"type":"string"},"driveUrl":{"type":"string","format":"uri","description":"Direct browser URL for opening the Drive file."},"name":{"type":"string"},"score":{"type":"number"},"semanticScore":{"type":"number"},"keywordScore":{"type":"number"},"metadataScore":{"type":"number"}}}},"grounding":{"type":"object"},"stale":{"type":"boolean"},"sessionId":{"type":"string","format":"uuid","nullable":true,"description":"Session ID if a session was used, null otherwise."}}}}}}}},"404":{"description":"Session not found (when sessionId is provided but invalid)."}}}},"/api/v1/ask/session":{"post":{"summary":"Create a new Ask conversation session.","tags":["Ask"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","maxLength":255,"description":"Optional display title for the session."},"projectId":{"type":"string","format":"uuid","description":"Optional project scope. All queries in this session default to this project."}}}}}},"responses":{"200":{"description":"Session created.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"}}}}}}}}}}},"/api/v1/ask/sessions":{"get":{"summary":"List all Ask conversation sessions.","tags":["Ask"],"responses":{"200":{"description":"List of sessions (without message history).","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"title":{"type":"string","nullable":true},"projectId":{"type":"string","format":"uuid","nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}}}}}}}}}}},"/api/v1/ask/session/{sessionId}":{"delete":{"summary":"Delete an Ask conversation session and all its history.","tags":["Ask"],"parameters":[{"in":"path","name":"sessionId","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Session deleted."},"404":{"description":"Session not found."}}}},"/api/v1/drive/health":{"get":{"summary":"Check auth state and confirm yantramate folder is visible","tags":["Drive"],"responses":{"200":{"description":"Health snapshot"}}}},"/api/v1/drive/auth/login":{"get":{"summary":"Redirect to Google's OAuth consent screen","tags":["Drive"],"responses":{"302":{"description":"Redirect to Google"}}}},"/api/v1/drive/auth/callback":{"get":{"summary":"OAuth callback. Exchanges code for tokens and persists them to disk.","tags":["Drive"],"parameters":[{"in":"query","name":"code","schema":{"type":"string"}},{"in":"query","name":"error","schema":{"type":"string"}}],"responses":{"200":{"description":"Tokens saved"},"400":{"description":"Missing code or OAuth error"}}}},"/api/v1/drive/projects":{"get":{"summary":"List top-level folders under yantramate (each = one project)","tags":["Drive"],"responses":{"200":{"description":"Project folder list","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"count":{"type":"integer"},"projects":{"type":"array","items":{"$ref":"#/components/schemas/DriveFile"}}}}}}}}}}}},"/api/v1/drive/projects/{projectId}/files":{"get":{"summary":"Recursively list all files in a project folder","tags":["Drive"],"parameters":[{"in":"path","name":"projectId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All files inside the project subtree"}}}},"/api/v1/drive/projects/by-name/{projectName}/files":{"get":{"summary":"Recursively list all files in a project folder, resolved by folder name","tags":["Drive"],"parameters":[{"in":"path","name":"projectName","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"All files inside the project subtree"},"404":{"description":"No top-level project folder with that name"}}}},"/api/v1/drive/files/{fileId}":{"get":{"summary":"Get metadata for a single file","tags":["Drive"],"parameters":[{"in":"path","name":"fileId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"File metadata","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/DriveFile"}}}}}}}}},"/api/v1/drive/files/{fileId}/content":{"get":{"summary":"Live-fetch the current content of a file from Drive","description":"Returns plain text for Google Docs/Sheets/Slides (exported on the fly) and base64 bytes for binary files.\nThis is the **freshness path** — every call hits Drive directly, never the cache/index.\n","tags":["Drive"],"parameters":[{"in":"path","name":"fileId","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Live file content","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/DriveFileContent"}}}}}}}}},"/api/v1/drive/sync/bootstrap":{"post":{"summary":"One-time full crawl of yantramate and capture initial Changes API cursor","tags":["Drive"],"responses":{"200":{"description":"Summary of files walked + new page token"}}}},"/api/v1/drive/sync/delta":{"post":{"summary":"Run a delta sync — return everything that changed since the stored cursor","tags":["Drive"],"responses":{"200":{"description":"Changes list + updated page token","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"changeCount":{"type":"integer"},"newPageToken":{"type":"string"},"changes":{"type":"array","items":{"$ref":"#/components/schemas/DriveChange"}}}}}}}}}}}},"/api/v1/drive/sync/status":{"get":{"summary":"Show the current Changes API cursor and last sync timestamp","tags":["Drive"],"responses":{"200":{"description":"Sync state"}}}},"/api/v1/drive/tree":{"get":{"summary":"Hierarchical view of the yantramate root annotated with index state","description":"Walks the Drive root subtree and joins it with what's in the vector DB.\nEach file node carries `indexed`, `chunkCount`, `supported`, and `lastIndexedAt`\nso you can immediately spot files Drive has but the index doesn't (e.g.\nunsupported MIME, extraction error, not yet ingested).\n","tags":["Drive"],"responses":{"200":{"description":"Annotated Drive tree + summary counters","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"summary":{"type":"object","properties":{"folderCount":{"type":"integer"},"fileCount":{"type":"integer"},"indexedCount":{"type":"integer"},"unsupportedCount":{"type":"integer"},"missingFromIndexCount":{"type":"integer"},"totalChunkCount":{"type":"integer"}}},"tree":{"type":"object"}}}}}}}}}}},"/api/v1/email/chat":{"post":{"summary":"Chat with the AI assistant over your indexed mailbox.","description":"Tool-calling chat backed by the locally-stored email index (Postgres + pgvector,\nkept fresh by the IMAP IDLE watcher). The assistant can list recent mail,\nsummarise time windows, semantic-search by topic, and pull a full body by id.\n\n**Sessions**: pass `sessionId` (from `POST /email/chat/session`) to use server-side\nconversation history. The server loads prior turns automatically and persists each\nnew exchange, so clients should not send chat history.\n","tags":["Email"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["message","sessionId"],"properties":{"message":{"type":"string","example":"What's new in my mail today?"},"sessionId":{"type":"string","format":"uuid","description":"UUID from POST /email/chat/session. Enables server-side history."}}}}}},"responses":{"200":{"description":"Assistant answer plus tool trace.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"object","properties":{"answer":{"type":"string"},"sessionId":{"type":"string","format":"uuid","nullable":true},"toolTrace":{"type":"array","items":{"type":"object","properties":{"name":{"type":"string"},"arguments":{"type":"object"},"result":{}}}},"indexStats":{"$ref":"#/components/schemas/EmailIndexStats"},"usage":{"type":"object","properties":{"promptTokens":{"type":"integer"},"completionTokens":{"type":"integer"},"totalTokens":{"type":"integer"}}}}}}}}}},"404":{"description":"Session not found."}}}},"/api/v1/email/chat/session":{"post":{"summary":"Create a new chat session","description":"Creates a new server-side chat session for the given account. Returns a `sessionId` UUID\nthat can be passed to `POST /email/chat` for persistent conversation history.\n","tags":["Email"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["accountUser"],"properties":{"accountUser":{"type":"string","description":"Email account whose indexed mailbox this session belongs to.","example":"you@example.com"},"title":{"type":"string","description":"Optional display title for the session.","example":"Daily digest check"}}}}}},"responses":{"200":{"description":"Session created","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"}}}}}}}}}}},"/api/v1/email/chat/sessions":{"post":{"summary":"List chat sessions for an account","description":"Returns all chat sessions for the resolved account, ordered by most recently updated.\nMessages are not included — use `POST /email/chat` with the sessionId to continue.\n","tags":["Email"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["accountUser"],"properties":{"accountUser":{"type":"string","description":"Email account whose sessions should be listed.","example":"you@example.com"}}}}}},"responses":{"200":{"description":"List of sessions","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/EmailChatSession"}}}}}}}}}},"/api/v1/email/chat/session/{sessionId}":{"delete":{"summary":"Delete a chat session","description":"Permanently deletes the chat session and all its stored messages.","tags":["Email"],"parameters":[{"in":"path","name":"sessionId","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Session deleted"},"404":{"description":"Session not found"}}}},"/api/v1/email/list":{"get":{"summary":"List emails with Gmail-style filters (sender, subject, date range, seen, flagged, has-attachment, full-text)","description":"Returns recent emails newest-first, with optional filtering and inclusion of body / attachment metadata.\nThis endpoint is query-param based so it can be tested directly from Swagger.\n\n**Filtering** uses IMAP server-side SEARCH where possible (fast). The `hasAttachment` filter\nrequires fetching body structures and is heavier. Combine it with other filters to narrow first.\n\n**Pagination** uses `page` (1-based) and `limit` over the filtered UID set.\n\n**Body / attachments** are *not* fetched by default. Opt in with `includeBody` and/or `includeAttachments`.\nFor attachment content (base64), use `GET /api/v1/email/get`.\n\nA JSON-body POST variant is also accepted at the same path for clients that should not put IMAP credentials\nin a URL.\n","tags":["Email"],"parameters":[{"in":"query","name":"host","required":true,"schema":{"type":"string","example":"imap.titan.email"}},{"in":"query","name":"port","schema":{"type":"integer","default":993,"minimum":1,"maximum":65535}},{"in":"query","name":"user","required":true,"schema":{"type":"string","example":"you@example.com"}},{"in":"query","name":"password","required":true,"schema":{"type":"string"}},{"in":"query","name":"secure","schema":{"type":"boolean","default":true}},{"in":"query","name":"mailbox","schema":{"type":"string","default":"INBOX","example":"INBOX"}},{"in":"query","name":"limit","schema":{"type":"integer","default":20,"minimum":1,"maximum":200}},{"in":"query","name":"page","schema":{"type":"integer","default":1,"minimum":1}},{"in":"query","name":"search","description":"Free-text search across subject and body.","schema":{"type":"string","example":"invoice"}},{"in":"query","name":"from","description":"Sender substring match.","schema":{"type":"string","example":"stripe.com"}},{"in":"query","name":"to","description":"Recipient substring match.","schema":{"type":"string"}},{"in":"query","name":"subject","description":"Subject substring match.","schema":{"type":"string","example":"receipt"}},{"in":"query","name":"since","description":"Only emails on/after this date. Accepts ISO date or datetime.","schema":{"type":"string","example":"2026-05-01"}},{"in":"query","name":"before","description":"Only emails before this date. Accepts ISO date or datetime.","schema":{"type":"string","example":"2026-05-22"}},{"in":"query","name":"seen","description":"true = read only, false = unread only, omit = both.","schema":{"type":"boolean"}},{"in":"query","name":"flagged","description":"true = starred/flagged only, false = unflagged only.","schema":{"type":"boolean"}},{"in":"query","name":"hasAttachment","description":"true = only emails with attachments, false = only without.","schema":{"type":"boolean"}},{"in":"query","name":"includeBody","description":"Include parsed text + HTML body in each item.","schema":{"type":"boolean","default":false}},{"in":"query","name":"includeAttachments","description":"Include attachment metadata. Content not included.","schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"Filtered, paginated list of messages (newest UID first).","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"total":{"type":"integer"},"page":{"type":"integer"},"limit":{"type":"integer"},"pages":{"type":"integer"},"messages":{"type":"array","items":{"$ref":"#/components/schemas/EmailListItem"}}}}}}}}},"400":{"description":"Validation error"},"500":{"description":"IMAP connection or fetch failed"}}}},"/api/v1/email/get":{"get":{"summary":"Fetch one email by UID with full body (text + html) and attachments","description":"Returns the complete parsed email: text body, HTML body, headers (cc/bcc/reply-to/message-id),\nand attachment metadata. Set `includeAttachmentContent=true` to also receive each attachment's\nbase64-encoded content (useful for download / preview).\n\nA JSON-body POST variant is also accepted at the same path for clients that should not put IMAP credentials\nin a URL.\n","tags":["Email"],"parameters":[{"in":"query","name":"host","required":true,"schema":{"type":"string","example":"imap.titan.email"}},{"in":"query","name":"port","schema":{"type":"integer","default":993,"minimum":1,"maximum":65535}},{"in":"query","name":"user","required":true,"schema":{"type":"string","example":"you@example.com"}},{"in":"query","name":"password","required":true,"schema":{"type":"string"}},{"in":"query","name":"secure","schema":{"type":"boolean","default":true}},{"in":"query","name":"mailbox","schema":{"type":"string","default":"INBOX"}},{"in":"query","name":"uid","required":true,"schema":{"type":"integer","example":1234}},{"in":"query","name":"includeAttachmentContent","description":"Include base64 content for each attachment.","schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"Full email detail.","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/EmailDetail"}}}}}},"400":{"description":"Validation error"},"404":{"description":"Email not found at that UID in the mailbox"},"500":{"description":"IMAP connection or fetch failed"}}}},"/api/v1/email/sync":{"post":{"summary":"Manually sync the mailbox to fetch missed emails","description":"Connects to the IMAP server and ingests any emails that are not currently stored in the vector database.\nUseful when the background IDLE watcher was disconnected or missed events.\n\n**Use `connectionId`** — the UUID returned by `POST /email/credentials`. The background IDLE listener\nhandles new mail automatically; call this endpoint only to backfill missed emails.\n\nIf `ingested: 0` and `totalPending: 0`, the database is already fully in sync.\n","tags":["Email"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["connectionId"],"properties":{"connectionId":{"type":"string","format":"uuid","description":"UUID returned by POST /email/credentials. Use this instead of raw credentials.","example":"e3cdb961-f790-4a69-93d8-5ed74f346544"},"mailbox":{"type":"string","default":"INBOX","example":"INBOX"},"limit":{"type":"integer","default":50,"maximum":1000,"description":"Max emails to ingest in this batch. Increase to sync more at once."}}}}}},"responses":{"200":{"description":"Sync completed","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"ingested":{"type":"integer","description":"Number of new emails ingested in this batch"},"totalPending":{"type":"integer","description":"Total emails missing before this sync ran"}}}}}}}}}}},"/api/v1/email/credentials":{"post":{"summary":"Save an IMAP credential securely","description":"Saves IMAP connection details in the database and returns a UUID `connectionId`.\nThis UUID can be used in the `POST /email/sync` and `POST /email/chat` endpoints.\nThe backend will immediately hot-start a background IDLE listener for this account.\n\n**Duplicate check**: If the same email address is already registered, returns `409 Conflict`\nwithout re-verifying credentials or re-embedding.\n","tags":["Email"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ImapCreds"}}}},"responses":{"200":{"description":"Credentials saved","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":"The connectionId UUID"}}}}}}}},"409":{"description":"Email already registered","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Email you@example.com is already registered. Remove it first before re-adding."}}}}}}}},"get":{"summary":"List all saved email credentials","description":"Returns all registered IMAP credentials. Passwords are stripped from the response.\n","tags":["Email"],"responses":{"200":{"description":"List of saved credentials","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"host":{"type":"string"},"port":{"type":"integer"},"user":{"type":"string"},"label":{"type":"string","nullable":true},"secure":{"type":"boolean"},"createdAt":{"type":"string","format":"date-time"}}}}}}}}}}}},"/api/v1/email/credentials/{emailAddress}":{"delete":{"summary":"Delete a registered email by its address","description":"Removes the IMAP credential record for the given email address.\n","tags":["Email"],"parameters":[{"in":"path","name":"emailAddress","required":true,"schema":{"type":"string","example":"you@example.com"},"description":"The email address (IMAP user) to unregister."}],"responses":{"200":{"description":"Credential deleted","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"}}}}}},"404":{"description":"No credential found for that email address"}}}},"/api/v1/projects":{"get":{"summary":"List all projects mirrored from Drive","tags":["Projects"],"responses":{"200":{"description":"Project list"}}}},"/api/v1/projects/{id}/files":{"get":{"summary":"List indexed files for a project","tags":["Projects"],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"Files indexed under the project"}}}}},"components":{"schemas":{"DriveFile":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"mimeType":{"type":"string"},"parents":{"type":"array","items":{"type":"string"}},"modifiedTime":{"type":"string","format":"date-time","nullable":true},"headRevisionId":{"type":"string","nullable":true},"size":{"type":"integer","nullable":true},"isFolder":{"type":"boolean"},"trashed":{"type":"boolean"}}},"DriveFileContent":{"type":"object","properties":{"fileId":{"type":"string"},"name":{"type":"string"},"mimeType":{"type":"string"},"encoding":{"type":"string","enum":["text","binary"]},"content":{"type":"string","description":"Plain text for Docs/Sheets/Slides. Base64 for binary files (PDF, image, docx)."},"bytes":{"type":"integer"}}},"DriveChange":{"type":"object","properties":{"fileId":{"type":"string"},"removed":{"type":"boolean"},"file":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/DriveFile"}]}}},"ImapCreds":{"type":"object","required":["host","user","password"],"properties":{"host":{"type":"string","example":"imap.titan.email"},"port":{"type":"integer","default":993},"user":{"type":"string","example":"you@example.com"},"password":{"type":"string"},"secure":{"type":"boolean","default":true},"label":{"type":"string","nullable":true,"description":"Optional display label for a saved credential.","example":"Work Email"}}},"EmailIndexStats":{"type":"object","properties":{"accountUser":{"type":"string"},"totalEmails":{"type":"integer"},"totalChunks":{"type":"integer"},"mailboxes":{"type":"array","items":{"type":"string"}},"oldestDate":{"type":"string","format":"date-time","nullable":true},"newestDate":{"type":"string","format":"date-time","nullable":true},"lastIndexedAt":{"type":"string","format":"date-time","nullable":true}}},"EmailChatSession":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"accountUser":{"type":"string"},"title":{"type":"string","nullable":true},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"EmailAttachment":{"type":"object","properties":{"filename":{"type":"string","nullable":true},"contentType":{"type":"string","example":"application/pdf"},"size":{"type":"integer","example":48213},"contentId":{"type":"string","nullable":true},"content":{"type":"string","description":"Base64-encoded content. Only present when requested."}}},"EmailListItem":{"type":"object","properties":{"uid":{"type":"integer","example":1234},"seq":{"type":"integer","example":87},"date":{"type":"string","format":"date-time","nullable":true},"subject":{"type":"string","nullable":true},"from":{"type":"string","nullable":true},"to":{"type":"string","nullable":true},"cc":{"type":"string","nullable":true},"flags":{"type":"array","items":{"type":"string"},"example":["\\Seen"]},"hasAttachments":{"type":"boolean"},"bodyText":{"type":"string","description":"Only present when includeBody=true."},"bodyHtml":{"type":"string","nullable":true,"description":"Only present when includeBody=true."},"attachments":{"type":"array","description":"Only present when includeAttachments=true (metadata, no content).","items":{"$ref":"#/components/schemas/EmailAttachment"}}}},"EmailDetail":{"type":"object","properties":{"uid":{"type":"integer"},"seq":{"type":"integer"},"date":{"type":"string","format":"date-time","nullable":true},"subject":{"type":"string","nullable":true},"from":{"type":"string","nullable":true},"to":{"type":"string","nullable":true},"cc":{"type":"string","nullable":true},"bcc":{"type":"string","nullable":true},"replyTo":{"type":"string","nullable":true},"messageId":{"type":"string","nullable":true},"inReplyTo":{"type":"string","nullable":true},"references":{"type":"array","items":{"type":"string"}},"flags":{"type":"array","items":{"type":"string"}},"bodyText":{"type":"string","nullable":true},"bodyHtml":{"type":"string","nullable":true},"attachments":{"type":"array","items":{"$ref":"#/components/schemas/EmailAttachment"}}}}}},"tags":[{"name":"Ask","description":"Question answering over indexed Drive content."},{"name":"Drive","description":"Google Drive read-only access. OAuth tokens are persisted locally so re-auth is not needed between restarts."},{"name":"Email","description":"IMAP email access, indexing, chat, sessions, and saved connection management."},{"name":"Projects","description":"Projects mirrored from top-level folders of the yantramate Drive root, plus their indexed files."}]}